Update the history plugin.
History deals with two tricky issues: autoplay and consistently setting
the currentTime as soon as possible.

To determine whether or not the track needs to play on initial load, the
player success callback is proxied since there doesn't seem to be a way
to specify multiple success callbacks. This should work whether or not
the player is initialized with $.mediaelementplayer() or $.cuePlaylist(),
and should account for the default autoplay attribute.

Setting the currentTime has to be done at certain points during the file
load process depending on which browser/OS is being used. iOS 7 doesn't
allow it to be set until the track is being buffered, which will cause
the track to play from the beginning until it's able to skip ahead.

Others need to wait until the 'loadedmetadata' or 'canplay' events.
Setting it during the initial 'playing' event seems to work in most
browsers.